refactor(compiler-cli): move listLazyRoutes to new private API
