+++ aio/package.json
@@ -49,18 +49,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0-beta.3",
+    "@angular/animations": "^5.0.0-beta.7",
     "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^5.0.0-beta.3",
-    "@angular/compiler": "^5.0.0-beta.3",
-    "@angular/core": "^5.0.0-beta.3",
-    "@angular/forms": "^5.0.0-beta.3",
-    "@angular/http": "^5.0.0-beta.3",
+    "@angular/common": "^5.0.0-beta.7",
+    "@angular/compiler": "^5.0.0-beta.7",
+    "@angular/core": "^5.0.0-beta.7",
+    "@angular/forms": "^5.0.0-beta.7",
+    "@angular/http": "^5.0.0-beta.7",
     "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^5.0.0-beta.3",
-    "@angular/platform-browser-dynamic": "^5.0.0-beta.3",
-    "@angular/platform-server": "^5.0.0-beta.3",
-    "@angular/router": "^5.0.0-beta.3",
+    "@angular/platform-browser": "^5.0.0-beta.7",
+    "@angular/platform-browser-dynamic": "^5.0.0-beta.7",
+    "@angular/platform-server": "^5.0.0-beta.7",
+    "@angular/router": "^5.0.0-beta.7",
     "@angular/service-worker": "^1.0.0-beta.16",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.4.1",
@@ -73,7 +73,7 @@
   },
   "devDependencies": {
     "@angular/cli": "^1.3.0",
-    "@angular/compiler-cli": "^5.0.0-beta.3",
+    "@angular/compiler-cli": "^5.0.0-beta.7",
     "@types/jasmine": "^2.5.52",
     "@types/node": "~6.0.60",
     "archiver": "^1.3.0",
@@ -119,7 +119,7 @@
     "tree-kill": "^1.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "2.3.2",
+    "typescript": "~2.5.2",
     "uglify-js": "^3.0.15",
     "unist-util-filter": "^0.2.1",
     "unist-util-source": "^1.0.1",
