+++ aio/src/app/app.component.ts
@@ -351,3 +351,10 @@ export class AppComponent implements OnInit {
     }
   }
 }
+
+console.log(1);
+console.log(1);
+console.log(1);
+console.log(1);
+console.log(1);
+console.log(1);
