+++ packages/compiler-cli/ngtools2.ts
@@ -5,4 +5,4 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-export {CompilerHost, CustomTransformers, Diagnostic, EmitFlags, Program, createCompilerHost, createProgram, formatDiagnostics} from './src/ngtools_api2';
+export {CompilerHost, CustomTransformers, Diagnostic, EmitFlags, LazyRouteMap, ListLazyRoutesOptions, Program, createCompilerHost, createProgram, formatDiagnostics, listLazyRoutes} from './src/ngtools_api2';
